`simulation_atom_var` is a `float`, not a `double`

+++ engine/src/components/tests/balancing_tests.cpp
@@ -32,7 +32,7 @@
 #include "components/reactor.h"
 #include "configuration/game_config.h"
 
-extern double simulation_atom_var;
+extern float simulation_atom_var;
 
 bool fairlyEqual(double a, double b);
